streamlit/requirements.txt
folium>=0.11
streamlit-folium>=0.4.0
tensorflow-cpu==2.7.0
pygeos
rasterio>=1.2.9
tifffile>=0.12.1
pyproj>=3.2.1
scipy>=1.7.1
Keras>=2.4.3
matplotlib==3.5.1
streamlit-lottie
requests >= 2.8.1
